dashboard/requirements.txt
streamlit
paho-mqtt
pandas
requests
flask
